trustpub/tokens/exchange: Rename `tests` module to `github_tests`
